third_party/blink/web_tests/external/wpt/dom/events/Event-dispatch-order-at-target-expected.txt

This is a testharness.js-based test.
FAIL Listeners are invoked in correct order (AT_TARGET phase) assert_array_equals: bubbles: true expected property 0 to be "capturing" but got "bubbling" (expected array ["capturing", "bubbling"] got ["bubbling", "capturing"])
Harness: the test ran to completion.
